Sample simulated data for exposure-response models with covariates

A synthetic dataset bundled with the package and used throughout its documentation and examples, with response columns illustrating each of erglm's supported glm() families.

Usage

erglm_data

Format

A data frame with columns:

id

Identifier

sex

Sex

age

Age

weight

Weight

dose

Nominal dose, units not specified

treatment

Treatment

aucss

AUCss

cmaxss

Cmax,ss

ae1

Binary response 1 value (for binomial models)

ae2

Binary response 2 value (for binomial models)

ae_count

Count response (for poisson models)

biomarker_change

Continuous response, can be negative (for gaussian models)

ae_duration

Continuous, strictly positive, right-skewed response (for gamma models)

Details

This simulated dataset is entirely synthetic. See the package source for the data-generating code.

Examples

erglm_data
#>      id    sex age weight dose treatment   aucss cmaxss ae1 ae2 ae_count
#> 1     1   Male  35     79  200      Drug  673.09  97.33   0   1        1
#> 2     2 Female  22     58  200      Drug 2806.12 300.62   1   1        6
#> 3     3 Female  28     58    0   Placebo    0.00   0.00   0   0        1
#> 4     4 Female  18     57  100      Drug 1169.04 197.78   1   1        0
#> 5     5   Male  28     77  100      Drug  377.29  51.43   0   0        0
#> 6     6 Female  19     76  200      Drug  327.08  25.37   1   0        0
#> 7     7   Male  30     70    0   Placebo    0.00   0.00   0   0        0
#> 8     8 Female  34     60  100      Drug 1207.64 133.02   1   1        1
#> 9     9   Male  21     89    0   Placebo    0.00   0.00   0   0        0
#> 10   10 Female  34     56  200      Drug  253.59  30.99   0   0        1
#> 11   11   Male  26     90  100      Drug 3145.29 257.18   1   1       14
#> 12   12 Female  30     59  200      Drug  404.70  56.14   1   1        0
#> 13   13 Female  27     66  200      Drug 1401.64 140.89   1   1        1
#> 14   14 Female  18     65    0   Placebo    0.00   0.00   0   0        0
#> 15   15   Male  27     61  200      Drug 1362.37 113.83   1   0        1
#> 16   16   Male  24     73  100      Drug  458.44  88.07   1   0        0
#> 17   17   Male  24     71  200      Drug  582.98  43.99   0   0        0
#> 18   18   Male  29     67    0   Placebo    0.00   0.00   0   0        0
#> 19   19   Male  30     68  200      Drug  385.85  37.41   0   0        1
#> 20   20   Male  27     90    0   Placebo    0.00   0.00   0   0        0
#> 21   21 Female  18     78    0   Placebo    0.00   0.00   0   0        0
#> 22   22   Male  35     69  100      Drug 1619.99 178.41   1   1        0
#> 23   23   Male  33     80    0   Placebo    0.00   0.00   0   0        0
#> 24   24 Female  33     62    0   Placebo    0.00   0.00   0   0        1
#> 25   25   Male  25     65  100      Drug 1130.99 116.01   1   0        1
#> 26   26 Female  18     66  100      Drug 1226.54  90.10   1   1        1
#> 27   27   Male  24     77    0   Placebo    0.00   0.00   0   0        0
#> 28   28 Female  25     78    0   Placebo    0.00   0.00   0   0        1
#> 29   29 Female  22     60  200      Drug 1461.41 155.93   1   1        1
#> 30   30 Female  18     55  200      Drug 3255.73 496.64   1   1       14
#> 31   31 Female  30     59  100      Drug  833.21  60.11   1   0        1
#> 32   32   Male  22     72  200      Drug  579.33  57.58   0   0        1
#> 33   33 Female  20     67  200      Drug  428.25  45.68   0   0        2
#> 34   34 Female  18     60  200      Drug 1060.98 114.53   1   1        2
#> 35   35 Female  33     51  100      Drug 1112.25  77.93   1   1        2
#> 36   36   Male  27     72    0   Placebo    0.00   0.00   1   0        1
#> 37   37 Female  25     72  200      Drug  274.34  22.59   1   0        1
#> 38   38 Female  18     65  100      Drug 1364.74  84.57   1   1        1
#> 39   39 Female  20     62  200      Drug  385.89  47.78   0   1        0
#> 40   40   Male  22     64  100      Drug 1455.61 121.91   1   0        3
#> 41   41 Female  28     62    0   Placebo    0.00   0.00   0   0        0
#> 42   42 Female  35     79  200      Drug 2355.54 280.81   1   1        8
#> 43   43   Male  31     81  200      Drug 1122.65 184.41   1   1        2
#> 44   44   Male  27     73  100      Drug  736.41  80.00   1   0        1
#> 45   45 Female  28     63  200      Drug  584.08  40.82   0   0        0
#> 46   46 Female  30     63  100      Drug  869.92  71.52   1   1        2
#> 47   47 Female  32     79  100      Drug 1317.13  60.84   1   1        1
#> 48   48 Female  19     51  100      Drug  383.24  16.20   1   1        0
#> 49   49   Male  24     65    0   Placebo    0.00   0.00   0   0        1
#> 50   50 Female  29     60  200      Drug 2681.92 240.82   1   1        5
#> 51   51 Female  25     64  200      Drug 1887.84 242.10   1   1        3
#> 52   52   Male  27     61    0   Placebo    0.00   0.00   0   0        0
#> 53   53   Male  32     58    0   Placebo    0.00   0.00   0   0        2
#> 54   54 Female  18     74  100      Drug  795.49  66.98   1   1        1
#> 55   55   Male  19     60    0   Placebo    0.00   0.00   0   0        0
#> 56   56   Male  30     77  200      Drug 3722.99 288.25   1   1       12
#> 57   57 Female  31     61  100      Drug  458.39  44.61   1   0        2
#> 58   58   Male  22     75  200      Drug 3052.06 167.61   1   1        9
#> 59   59   Male  28     78  100      Drug  480.92  77.50   1   0        1
#> 60   60 Female  20     57    0   Placebo    0.00   0.00   1   0        2
#> 61   61   Male  27     71  200      Drug 2170.83 262.77   1   1        1
#> 62   62 Female  35     49  100      Drug  789.62  60.10   1   1        1
#> 63   63   Male  18     65  200      Drug 3642.90 373.90   1   1       20
#> 64   64   Male  19     73  100      Drug  354.18  35.21   0   0        1
#> 65   65   Male  27     68    0   Placebo    0.00   0.00   0   0        0
#> 66   66   Male  26     84    0   Placebo    0.00   0.00   0   0        0
#> 67   67 Female  24     64  100      Drug  877.32  96.26   1   0        0
#> 68   68 Female  33     55  100      Drug  525.98  30.49   0   0        1
#> 69   69 Female  28     80    0   Placebo    0.00   0.00   0   0        1
#> 70   70 Female  22     62  200      Drug 1044.01 150.89   1   0        0
#> 71   71   Male  27     72    0   Placebo    0.00   0.00   0   0        1
#> 72   72   Male  21     58  200      Drug 1767.37 186.44   1   1        1
#> 73   73 Female  19     57  100      Drug 1044.31 118.10   1   1        1
#> 74   74 Female  20     79  100      Drug 2416.47 185.79   1   1        4
#> 75   75 Female  25     67  100      Drug  226.23  17.55   0   0        1
#> 76   76   Male  35     94  200      Drug 2015.46 115.51   1   1        3
#> 77   77   Male  24     56    0   Placebo    0.00   0.00   0   0        1
#> 78   78   Male  31     56    0   Placebo    0.00   0.00   0   0        0
#> 79   79   Male  25     88  100      Drug  974.22 147.32   1   1        0
#> 80   80   Male  27     76  100      Drug 1808.76 177.28   1   1        3
#> 81   81   Male  21     82  200      Drug  768.96  56.50   1   1        2
#> 82   82 Female  29     64  200      Drug  597.49  86.14   1   1        2
#> 83   83   Male  31     67  200      Drug 1849.66 129.64   1   1        2
#> 84   84   Male  30     82    0   Placebo    0.00   0.00   0   0        0
#> 85   85 Female  31     75    0   Placebo    0.00   0.00   0   0        0
#> 86   86 Female  30     67    0   Placebo    0.00   0.00   0   0        1
#> 87   87 Female  31     55    0   Placebo    0.00   0.00   1   0        1
#> 88   88   Male  29     76  200      Drug 1217.92  93.03   1   1        1
#> 89   89   Male  34     69    0   Placebo    0.00   0.00   0   1        0
#> 90   90   Male  30     68  100      Drug  702.45  30.62   1   0        1
#> 91   91 Female  20     58  100      Drug 1302.04 208.09   1   1        4
#> 92   92 Female  30     53  200      Drug  446.25  51.12   1   0        0
#> 93   93 Female  21     58  100      Drug  257.87  25.63   1   0        1
#> 94   94   Male  27     74  100      Drug 2246.04 201.54   1   1        4
#> 95   95   Male  26     66    0   Placebo    0.00   0.00   0   0        0
#> 96   96   Male  27     57  200      Drug  294.72  22.51   1   0        1
#> 97   97 Female  35     59    0   Placebo    0.00   0.00   0   0        1
#> 98   98   Male  28     63    0   Placebo    0.00   0.00   0   0        1
#> 99   99 Female  27     56    0   Placebo    0.00   0.00   1   0        1
#> 100 100   Male  26     66  200      Drug  441.22  53.43   1   0        0
#> 101 101   Male  29     77  100      Drug  483.02  40.05   1   0        1
#> 102 102   Male  20     91  200      Drug 1807.90 213.18   1   1        3
#> 103 103 Female  35     68  200      Drug 1408.28  87.97   1   0        0
#> 104 104 Female  32     72  100      Drug  793.38 180.39   1   0        1
#> 105 105 Female  18     47  200      Drug  818.91  45.14   1   1        0
#> 106 106 Female  29     51  100      Drug  162.86  13.09   0   0        1
#> 107 107 Female  30     50  200      Drug 3213.40 167.70   1   1       12
#> 108 108 Female  20     61    0   Placebo    0.00   0.00   0   0        1
#> 109 109   Male  19     83  100      Drug  890.56  65.19   1   0        1
#> 110 110   Male  22     73  100      Drug 2924.26 223.05   1   1       10
#> 111 111 Female  33     63    0   Placebo    0.00   0.00   1   0        0
#> 112 112   Male  22     72  200      Drug 2539.48 204.58   1   1        5
#> 113 113   Male  24     68  200      Drug  730.70  97.33   1   0        1
#> 114 114   Male  25     76  200      Drug  263.59  41.75   0   0        0
#> 115 115   Male  26     72  100      Drug 1425.73 177.81   1   1        2
#> 116 116   Male  23     58    0   Placebo    0.00   0.00   0   0        1
#> 117 117 Female  26     56    0   Placebo    0.00   0.00   0   1        1
#> 118 118   Male  28     73  100      Drug  461.75  31.40   1   0        0
#> 119 119 Female  25     63  200      Drug 2291.94  97.11   1   1        5
#> 120 120 Female  31     50  200      Drug  367.06  27.66   1   1        1
#> 121 121 Female  18     59  200      Drug  339.20  52.29   1   1        1
#> 122 122 Female  32     63    0   Placebo    0.00   0.00   0   0        0
#> 123 123 Female  21     64  100      Drug 2043.24 249.90   1   1        1
#> 124 124 Female  24     56  200      Drug  374.90  18.51   1   0        0
#> 125 125 Female  34     71    0   Placebo    0.00   0.00   0   1        0
#> 126 126   Male  29     76    0   Placebo    0.00   0.00   0   0        1
#> 127 127 Female  21     60  100      Drug  471.19  40.50   0   1        0
#> 128 128 Female  27     63  100      Drug 2950.71 274.01   1   1        9
#> 129 129 Female  31     63  200      Drug 1162.90 130.85   1   1        2
#> 130 130 Female  26     50  100      Drug  530.79  40.70   0   0        0
#> 131 131   Male  21     80  200      Drug 1007.34  97.58   1   0        1
#> 132 132   Male  21     68  100      Drug 1581.00 141.10   1   1        2
#> 133 133   Male  34     75  100      Drug 1157.31 203.29   1   0        0
#> 134 134   Male  32     85  100      Drug  224.63  21.84   1   0        1
#> 135 135   Male  31     59    0   Placebo    0.00   0.00   0   0        0
#> 136 136   Male  26     95  200      Drug  246.67  24.68   0   0        0
#> 137 137 Female  19     59  200      Drug  645.60  98.96   1   1        1
#> 138 138 Female  24     65    0   Placebo    0.00   0.00   0   0        0
#> 139 139   Male  19     68  100      Drug 1100.36  62.23   1   1        1
#> 140 140 Female  25     67    0   Placebo    0.00   0.00   1   0        1
#> 141 141   Male  35     69  200      Drug  628.59  97.99   0   0        1
#> 142 142 Female  21     72    0   Placebo    0.00   0.00   0   0        0
#> 143 143 Female  28     66  100      Drug  289.18  47.88   0   0        3
#> 144 144 Female  18     56    0   Placebo    0.00   0.00   0   0        0
#> 145 145   Male  18     64  200      Drug 1208.34 106.93   1   0        2
#> 146 146 Female  21     61  200      Drug 2775.36 498.60   1   1        7
#> 147 147 Female  28     72  100      Drug 2865.49 271.86   1   1        6
#> 148 148 Female  24     63  200      Drug 1677.63  90.13   1   1        3
#> 149 149   Male  25     61    0   Placebo    0.00   0.00   0   1        1
#> 150 150   Male  19     68  100      Drug 2051.81 148.59   1   1        3
#> 151 151   Male  28     73    0   Placebo    0.00   0.00   0   0        1
#> 152 152 Female  34     54  200      Drug  568.10  43.46   1   0        1
#> 153 153 Female  25     66    0   Placebo    0.00   0.00   0   1        0
#> 154 154   Male  32     61  200      Drug 1064.01 143.38   1   0        0
#> 155 155 Female  21     70  100      Drug  935.13 100.01   1   0        1
#> 156 156   Male  25     63  200      Drug  887.80  71.80   1   1        0
#> 157 157   Male  21     78  100      Drug  609.15 131.32   1   1        1
#> 158 158   Male  29     87  200      Drug 2547.68 359.10   1   1        5
#> 159 159   Male  30     65  200      Drug  560.82  89.20   1   0        1
#> 160 160   Male  28     75  100      Drug  341.38  23.41   0   1        0
#> 161 161 Female  31     51    0   Placebo    0.00   0.00   0   0        0
#> 162 162   Male  23     75    0   Placebo    0.00   0.00   0   0        0
#> 163 163   Male  35     79  100      Drug  674.44  73.28   1   0        0
#> 164 164   Male  26     64  100      Drug 1920.37 212.53   1   1        3
#> 165 165   Male  22     67  200      Drug  191.70  12.44   1   0        0
#> 166 166 Female  20     66  100      Drug  382.55  55.30   0   1        0
#> 167 167 Female  29     57    0   Placebo    0.00   0.00   1   0        0
#> 168 168   Male  25     77  100      Drug  339.44  34.73   1   1        0
#> 169 169 Female  28     60    0   Placebo    0.00   0.00   0   0        0
#> 170 170 Female  19     68    0   Placebo    0.00   0.00   0   1        0
#> 171 171 Female  32     60  200      Drug 1520.07 120.53   1   1        1
#> 172 172 Female  24     75  200      Drug  339.04  29.59   0   0        0
#> 173 173 Female  33     62  100      Drug  893.45 123.08   1   1        1
#> 174 174 Female  33     65  100      Drug  212.55  19.16   0   0        0
#> 175 175 Female  25     54    0   Placebo    0.00   0.00   0   0        1
#> 176 176 Female  24     68  200      Drug 1365.61 104.78   1   1        4
#> 177 177   Male  27     77  100      Drug  866.95 121.07   0   1        2
#> 178 178   Male  24     73  100      Drug 2454.94 197.49   1   1        6
#> 179 179 Female  28     62  200      Drug  497.48  48.54   1   0        0
#> 180 180   Male  24     68  200      Drug 1625.45  88.95   1   1        1
#> 181 181   Male  29     82  100      Drug 2504.67 185.93   1   1        7
#> 182 182   Male  33     74  200      Drug  818.65  87.23   1   1        0
#> 183 183   Male  30     62  100      Drug 1468.70 223.97   1   1        3
#> 184 184 Female  33     56    0   Placebo    0.00   0.00   0   0        0
#> 185 185   Male  31     82  200      Drug  503.62  78.53   1   0        1
#> 186 186 Female  20     60    0   Placebo    0.00   0.00   1   0        2
#> 187 187   Male  19     64    0   Placebo    0.00   0.00   1   0        0
#> 188 188   Male  35     69    0   Placebo    0.00   0.00   0   0        0
#> 189 189 Female  23     48  200      Drug 2280.14  92.72   1   1        4
#> 190 190 Female  22     58  100      Drug  867.80  69.57   1   0        0
#> 191 191 Female  26     66  100      Drug 1094.45 187.51   1   1        0
#> 192 192   Male  20     69  100      Drug  810.10  80.56   1   0        0
#> 193 193   Male  22     57  100      Drug  983.89 125.00   1   0        1
#> 194 194 Female  32     72  200      Drug 3255.37 390.22   1   1       17
#> 195 195 Female  20     56  100      Drug 1443.56 114.09   1   1        0
#> 196 196 Female  25     69    0   Placebo    0.00   0.00   0   0        1
#> 197 197 Female  26     69  100      Drug  540.91  78.51   1   0        0
#> 198 198   Male  22     75    0   Placebo    0.00   0.00   0   0        0
#> 199 199   Male  24     56  200      Drug  856.34  61.14   1   0        2
#> 200 200   Male  28     70  200      Drug  799.22  82.09   1   0        0
#> 201 201 Female  21     61  200      Drug 1753.61  94.00   1   1        4
#> 202 202 Female  28     61  100      Drug 2603.92 141.58   1   1        3
#> 203 203   Male  31     57  200      Drug 2577.04 185.83   1   0        4
#> 204 204 Female  20     68  200      Drug  393.77  49.72   1   0        2
#> 205 205   Male  29     83  200      Drug 2065.23 453.34   1   1        0
#> 206 206 Female  34     56    0   Placebo    0.00   0.00   1   0        1
#> 207 207 Female  30     51    0   Placebo    0.00   0.00   0   0        0
#> 208 208   Male  19     58  200      Drug 1724.40 112.84   1   0        1
#> 209 209 Female  27     58    0   Placebo    0.00   0.00   0   1        2
#> 210 210   Male  34     67    0   Placebo    0.00   0.00   0   0        0
#> 211 211 Female  21     51  200      Drug 1542.17 177.61   1   0        1
#> 212 212 Female  27     62  100      Drug  516.93  36.47   1   0        0
#> 213 213   Male  34     70    0   Placebo    0.00   0.00   0   0        0
#> 214 214 Female  32     75  100      Drug  726.04 136.93   1   0        2
#> 215 215 Female  34     50  200      Drug  670.34  77.26   1   0        0
#> 216 216 Female  24     80    0   Placebo    0.00   0.00   0   0        0
#> 217 217   Male  24     70    0   Placebo    0.00   0.00   0   0        0
#> 218 218   Male  30     70  200      Drug  648.47  68.27   1   0        0
#> 219 219 Female  29     61  100      Drug 1081.96 100.87   1   0        0
#> 220 220   Male  34     72  100      Drug  171.23  16.80   0   0        0
#> 221 221   Male  28     90  200      Drug 1901.77 110.02   1   1        3
#> 222 222   Male  24     57    0   Placebo    0.00   0.00   0   0        0
#> 223 223 Female  27     73  200      Drug 1842.74 306.78   1   1        3
#> 224 224 Female  31     73    0   Placebo    0.00   0.00   0   0        0
#> 225 225   Male  19     91  200      Drug  603.23  45.74   1   1        1
#> 226 226 Female  35     62    0   Placebo    0.00   0.00   1   0        0
#> 227 227 Female  33     64  200      Drug 1966.17 303.53   1   1        2
#> 228 228 Female  34     78    0   Placebo    0.00   0.00   0   0        1
#> 229 229 Female  23     50    0   Placebo    0.00   0.00   1   0        0
#> 230 230   Male  33     61  100      Drug 1218.94  73.69   1   1        0
#> 231 231 Female  21     64  100      Drug  234.20  23.74   0   0        1
#> 232 232   Male  27     85  100      Drug  496.35  68.39   1   0        0
#> 233 233   Male  24     79  200      Drug  303.59  25.24   1   0        1
#> 234 234 Female  23     59  100      Drug  675.52  51.05   1   0        1
#> 235 235   Male  18     79    0   Placebo    0.00   0.00   0   0        0
#> 236 236 Female  27     71  200      Drug  331.59  32.55   1   0        1
#> 237 237   Male  18     63  100      Drug 2273.11 135.72   1   1        3
#> 238 238   Male  18     69    0   Placebo    0.00   0.00   1   0        1
#> 239 239   Male  35     91  200      Drug 1789.80 180.54   1   1        1
#> 240 240   Male  26     72  100      Drug 1840.38 170.89   1   1        3
#> 241 241   Male  22     63  100      Drug  594.21  54.35   1   0        0
#> 242 242 Female  35     58  200      Drug 1338.90 147.45   1   1        0
#> 243 243   Male  21     74    0   Placebo    0.00   0.00   1   0        0
#> 244 244 Female  30     67  100      Drug  352.78  36.61   0   0        1
#> 245 245   Male  26     66  100      Drug  228.97  30.12   0   0        0
#> 246 246   Male  31     80    0   Placebo    0.00   0.00   0   1        1
#> 247 247   Male  30     80  200      Drug 3055.03 238.68   1   1        5
#> 248 248   Male  24     74  100      Drug 1012.23  58.33   1   1        2
#> 249 249   Male  27     85    0   Placebo    0.00   0.00   0   1        0
#> 250 250 Female  23     71    0   Placebo    0.00   0.00   0   0        0
#> 251 251 Female  30     49  200      Drug  601.15  60.65   1   1        0
#> 252 252   Male  26     62  100      Drug  372.56  46.87   1   0        1
#> 253 253   Male  21     86  100      Drug  326.80  15.26   1   0        0
#> 254 254   Male  18     77  200      Drug 2484.89 292.08   1   1        8
#> 255 255   Male  20     66  200      Drug  208.31  20.26   0   1        1
#> 256 256   Male  29     91    0   Placebo    0.00   0.00   0   0        0
#> 257 257   Male  18     78    0   Placebo    0.00   0.00   0   0        2
#> 258 258 Female  24     57    0   Placebo    0.00   0.00   0   0        0
#> 259 259   Male  24     81  100      Drug  286.87  39.78   0   0        0
#> 260 260 Female  32     71    0   Placebo    0.00   0.00   0   0        0
#> 261 261 Female  27     67  100      Drug  685.88  85.32   1   0        3
#> 262 262 Female  34     60  100      Drug  435.48  59.80   0   0        0
#> 263 263   Male  25     65  200      Drug  867.89  51.08   1   0        0
#> 264 264   Male  33     80    0   Placebo    0.00   0.00   0   0        0
#> 265 265   Male  20     76    0   Placebo    0.00   0.00   0   0        0
#> 266 266 Female  21     64  100      Drug 1048.37  67.56   1   0        1
#> 267 267 Female  21     56  200      Drug  446.83  41.10   1   0        0
#> 268 268 Female  30     59  200      Drug  833.28 110.67   1   0        3
#> 269 269 Female  26     76  100      Drug  884.21  78.18   1   0        3
#> 270 270   Male  21     83    0   Placebo    0.00   0.00   0   0        0
#> 271 271   Male  25     73    0   Placebo    0.00   0.00   0   0        1
#> 272 272   Male  21     66  200      Drug 2165.42 301.47   1   1        2
#> 273 273 Female  25     65  200      Drug 1121.83 117.53   1   0        1
#> 274 274 Female  31     67    0   Placebo    0.00   0.00   0   0        0
#> 275 275   Male  31     64  100      Drug  827.69  93.51   1   1        1
#> 276 276   Male  20     64  200      Drug 1265.56 154.55   1   1        1
#> 277 277   Male  18     59    0   Placebo    0.00   0.00   0   0        0
#> 278 278 Female  31     57  200      Drug  653.85 120.50   1   0        1
#> 279 279 Female  28     60  100      Drug  991.64  76.43   1   0        2
#> 280 280 Female  23     62  100      Drug  470.27  86.45   0   0        0
#> 281 281 Female  18     68    0   Placebo    0.00   0.00   0   0        0
#> 282 282 Female  33     74    0   Placebo    0.00   0.00   0   0        0
#> 283 283 Female  21     65  100      Drug 2490.56 242.82   1   1        6
#> 284 284 Female  27     58  200      Drug  273.96  16.93   1   0        0
#> 285 285   Male  18     79  100      Drug  376.13  55.44   1   1        0
#> 286 286 Female  28     70    0   Placebo    0.00   0.00   0   1        0
#> 287 287   Male  21     75  100      Drug 3106.73 360.53   1   1        8
#> 288 288   Male  22     74    0   Placebo    0.00   0.00   0   1        1
#> 289 289   Male  32     57    0   Placebo    0.00   0.00   0   0        1
#> 290 290   Male  24     79    0   Placebo    0.00   0.00   0   0        0
#> 291 291   Male  27     60    0   Placebo    0.00   0.00   1   0        0
#> 292 292 Female  29     65    0   Placebo    0.00   0.00   0   0        0
#> 293 293   Male  27     60  100      Drug  809.50  96.28   1   1        1
#> 294 294   Male  21     74  100      Drug 2584.89 502.81   1   1        1
#> 295 295 Female  31     56    0   Placebo    0.00   0.00   0   0        1
#> 296 296 Female  35     72    0   Placebo    0.00   0.00   0   1        0
#> 297 297 Female  25     66    0   Placebo    0.00   0.00   0   0        0
#> 298 298   Male  35     88    0   Placebo    0.00   0.00   0   0        0
#> 299 299   Male  25     70  200      Drug 1000.85 115.33   1   1        2
#> 300 300 Female  28     51  100      Drug  230.66  31.16   0   0        0
#>     biomarker_change ae_duration
#> 1               1.22       12.40
#> 2               4.87       13.70
#> 3              -1.83        5.26
#> 4               1.90        6.70
#> 5              -1.01        6.15
#> 6              -4.97       12.83
#> 7              -2.52        2.47
#> 8               1.58        9.13
#> 9              -0.76        9.20
#> 10             -1.22       10.63
#> 11              3.57       11.06
#> 12             -2.57        1.97
#> 13              1.28        4.30
#> 14             -1.24       12.11
#> 15              1.12       12.49
#> 16              1.38       15.56
#> 17             -0.72        2.56
#> 18             -3.67       10.06
#> 19             -4.46       17.89
#> 20             -1.02        4.40
#> 21             -2.14        5.21
#> 22              3.17        1.91
#> 23             -4.50        2.04
#> 24             -3.27        4.71
#> 25             -0.32        7.55
#> 26             -2.11        0.29
#> 27             -1.72        5.43
#> 28             -2.29        5.35
#> 29              1.09        4.17
#> 30              4.05        3.32
#> 31             -1.64        4.68
#> 32              0.25        8.29
#> 33             -1.56        8.34
#> 34             -0.72       22.87
#> 35              2.03        5.05
#> 36             -0.81        7.78
#> 37              0.13        9.31
#> 38              0.79       12.28
#> 39              1.37       17.77
#> 40              0.98        5.48
#> 41             -2.29        3.92
#> 42              2.42        7.81
#> 43             -0.89        7.70
#> 44              0.93        0.57
#> 45              1.33        4.17
#> 46             -0.26        3.83
#> 47              2.98        6.90
#> 48              0.13       14.92
#> 49             -3.04        4.33
#> 50              1.34        4.29
#> 51              0.68       10.65
#> 52             -1.65        0.85
#> 53             -1.18        1.21
#> 54             -2.33       12.03
#> 55              0.35       12.83
#> 56              5.05       28.20
#> 57              0.72       14.55
#> 58              3.40        7.29
#> 59             -0.47        8.16
#> 60             -1.33        5.17
#> 61              0.82       14.80
#> 62             -0.25        4.21
#> 63              4.71       24.12
#> 64             -2.31        7.34
#> 65             -4.43        7.89
#> 66             -3.62        5.85
#> 67              0.62       15.03
#> 68              0.56        6.11
#> 69             -0.42        7.93
#> 70              1.76        3.26
#> 71             -3.98        0.71
#> 72              1.26       14.65
#> 73              2.67        8.09
#> 74              2.24       17.13
#> 75             -1.88       29.83
#> 76              1.95        4.66
#> 77             -2.73       12.95
#> 78             -5.06        6.64
#> 79             -0.74       20.20
#> 80              2.76        7.58
#> 81              0.31        8.33
#> 82             -0.84       12.56
#> 83             -0.22        6.55
#> 84             -3.55        1.07
#> 85             -3.52        8.63
#> 86             -1.68       10.00
#> 87              0.07        6.85
#> 88              3.38       19.12
#> 89             -3.29        3.32
#> 90             -0.16        7.82
#> 91             -0.49       12.12
#> 92              0.05        9.49
#> 93              0.71        8.13
#> 94              1.37        4.08
#> 95             -4.46        1.58
#> 96              0.80        8.18
#> 97             -0.45        3.18
#> 98             -2.90        0.58
#> 99              1.05        5.87
#> 100             1.21       12.66
#> 101            -0.26        0.57
#> 102             0.86       15.06
#> 103             0.95       23.59
#> 104             0.55        2.62
#> 105             2.21       11.55
#> 106            -0.51        6.57
#> 107             2.05        8.89
#> 108            -2.71        4.66
#> 109             0.17        2.19
#> 110             3.48       19.44
#> 111            -3.22        2.58
#> 112             3.11        5.27
#> 113            -1.11        9.17
#> 114            -0.85        6.44
#> 115             0.19       13.45
#> 116             0.16        8.03
#> 117            -0.72        5.85
#> 118             0.02       10.30
#> 119             2.89       27.73
#> 120            -0.08        8.90
#> 121             2.22        2.37
#> 122             0.08        3.60
#> 123             2.30        8.83
#> 124             0.89       13.51
#> 125            -1.62        3.48
#> 126            -2.44        5.87
#> 127             0.13        4.36
#> 128             7.02        7.98
#> 129            -0.29       15.23
#> 130            -0.85       18.59
#> 131             0.98        6.53
#> 132             3.02       20.74
#> 133            -0.10        5.40
#> 134            -1.26        6.97
#> 135            -3.51        7.74
#> 136            -2.23        8.37
#> 137             0.08        4.70
#> 138            -0.70        8.43
#> 139             1.55        2.83
#> 140            -2.55        9.49
#> 141             2.20       15.36
#> 142            -0.44        1.98
#> 143             2.62       11.78
#> 144            -0.71        0.56
#> 145            -0.96        3.92
#> 146             4.10       10.25
#> 147             3.25        3.02
#> 148             1.01        7.10
#> 149            -3.19        9.28
#> 150             1.42        6.27
#> 151             0.70        0.90
#> 152             1.33        7.83
#> 153            -1.71        6.43
#> 154             1.99       25.90
#> 155             1.75        3.83
#> 156            -1.40        8.66
#> 157            -0.72        3.24
#> 158             6.22       10.19
#> 159             0.39       15.12
#> 160            -0.22        3.57
#> 161             0.10        3.94
#> 162            -0.91        2.56
#> 163             2.88        2.04
#> 164             1.72        2.12
#> 165            -0.33       13.45
#> 166             0.22        6.18
#> 167            -0.70        6.13
#> 168            -0.91        2.89
#> 169            -0.74        2.34
#> 170            -2.32        2.09
#> 171             1.35        7.26
#> 172            -1.11        3.98
#> 173             0.04       11.05
#> 174            -3.84       13.57
#> 175            -2.81        4.96
#> 176             0.20        9.16
#> 177             1.36       14.00
#> 178             3.51       14.78
#> 179            -2.43        6.56
#> 180             1.87       29.42
#> 181             0.05       18.48
#> 182            -0.19        6.97
#> 183            -1.51       14.36
#> 184             0.26        6.86
#> 185            -0.59        4.56
#> 186            -1.19        2.38
#> 187             0.44        3.74
#> 188            -2.01        6.35
#> 189             1.72       22.82
#> 190             0.86        1.31
#> 191             3.36       11.42
#> 192             0.37        9.56
#> 193             0.60        5.73
#> 194             3.19       25.35
#> 195             1.11       23.35
#> 196            -4.35       22.32
#> 197             0.10        2.48
#> 198            -0.85        8.14
#> 199            -0.21        4.23
#> 200             1.14       14.21
#> 201             3.98       15.51
#> 202             3.62        7.35
#> 203             4.54       16.55
#> 204            -0.12        5.30
#> 205             4.11       21.84
#> 206            -1.67       11.27
#> 207            -2.94        0.63
#> 208             0.32       24.21
#> 209             1.29        2.11
#> 210            -3.10        4.96
#> 211             0.91        8.26
#> 212            -2.43       12.24
#> 213            -4.78        2.21
#> 214             2.19        9.39
#> 215             1.02       25.58
#> 216            -2.89        6.76
#> 217            -3.43        2.04
#> 218            -3.35        1.97
#> 219             2.83       21.41
#> 220            -1.87        8.78
#> 221             1.10        6.89
#> 222            -3.21        2.81
#> 223             4.22        8.15
#> 224            -2.17        2.77
#> 225            -0.57        5.96
#> 226            -0.72        3.98
#> 227             2.51        3.16
#> 228            -3.30        4.36
#> 229            -3.12        3.57
#> 230             0.42       16.38
#> 231            -0.72       21.35
#> 232             0.26        7.14
#> 233            -1.79        5.93
#> 234            -0.74       30.85
#> 235            -1.23       17.73
#> 236             0.02        0.84
#> 237             1.14        4.48
#> 238            -2.24        2.44
#> 239            -1.50       12.06
#> 240             1.02       14.67
#> 241            -0.29        3.58
#> 242            -0.56        8.00
#> 243            -3.02        3.62
#> 244            -1.66        8.56
#> 245            -0.93       18.84
#> 246            -1.48       11.50
#> 247             2.40       28.51
#> 248            -0.45        3.95
#> 249            -1.59        2.33
#> 250            -0.97        7.89
#> 251            -0.52        5.23
#> 252            -2.49        0.60
#> 253            -1.00        5.56
#> 254             4.06       10.28
#> 255            -3.91       16.00
#> 256            -0.44        1.58
#> 257            -4.50        7.44
#> 258            -0.82        8.08
#> 259            -3.31        6.27
#> 260            -0.12        5.33
#> 261             0.54        6.61
#> 262             0.29        3.08
#> 263            -2.44       26.66
#> 264            -1.85        2.74
#> 265            -0.67        3.77
#> 266             0.19       14.64
#> 267            -0.90        2.20
#> 268            -0.32        2.35
#> 269             0.97        5.86
#> 270            -1.29        2.46
#> 271            -1.53        3.55
#> 272             2.46       16.10
#> 273             0.30        6.56
#> 274            -4.87        3.43
#> 275            -0.80       14.19
#> 276             1.58        9.24
#> 277            -2.72        1.97
#> 278            -2.96        7.70
#> 279            -1.49       11.70
#> 280            -4.61        2.15
#> 281            -2.87        8.75
#> 282            -0.72       13.88
#> 283             3.15       25.16
#> 284             1.72        7.35
#> 285            -1.54        7.30
#> 286            -4.07        5.78
#> 287             2.44        6.01
#> 288            -3.08        4.76
#> 289            -1.59        3.47
#> 290            -0.72        3.50
#> 291             0.81        3.23
#> 292            -1.23        8.64
#> 293            -0.84        2.03
#> 294             4.79        7.98
#> 295             1.10        4.59
#> 296            -2.54        3.22
#> 297             2.29        3.40
#> 298            -2.52       14.35
#> 299            -0.49       10.26
#> 300            -2.22        4.79