Sample simulated data for exposure-response models with covariates

A synthetic dataset bundled with the package and used throughout its documentation and examples, with response columns illustrating each of erglm's supported glm() families.

Usage

erglm_data

Format

A data frame with columns:

id

Identifier

sex

Sex

age

Age

weight

Weight

dose

Nominal dose, units not specified

treatment

Treatment

aucss

AUCss

cmaxss

Cmax,ss

ae1

Binary response 1 value (for binomial models)

ae2

Binary response 2 value (for binomial models)

ae_count

Count response (for poisson models)

biomarker_change

Continuous response, can be negative (for gaussian models)

ae_duration

Continuous, strictly positive, right-skewed response (for gamma models)

Details

This simulated dataset is entirely synthetic. See the package source for the data-generating code.

Examples

erglm_data
#>      id    sex age weight dose treatment    aucss  cmaxss ae1 ae2 ae_count
#> 1     1   Male  35     79  200      Drug  673.091  97.328   0   1        1
#> 2     2 Female  22     58  200      Drug 2806.115 300.615   1   1        6
#> 3     3 Female  28     58    0   Placebo    0.000   0.000   0   0        1
#> 4     4 Female  18     57  100      Drug 1169.045 197.783   1   1        0
#> 5     5   Male  28     77  100      Drug  377.288  51.429   0   0        0
#> 6     6 Female  19     76  200      Drug  327.079  25.373   1   0        0
#> 7     7   Male  30     70    0   Placebo    0.000   0.000   0   0        0
#> 8     8 Female  34     60  100      Drug 1207.639 133.024   1   1        1
#> 9     9   Male  21     89    0   Placebo    0.000   0.000   0   0        0
#> 10   10 Female  34     56  200      Drug  253.590  30.986   0   0        1
#> 11   11   Male  26     90  100      Drug 3145.293 257.177   1   1       14
#> 12   12 Female  30     59  200      Drug  404.697  56.141   1   1        0
#> 13   13 Female  27     66  200      Drug 1401.644 140.886   1   1        1
#> 14   14 Female  18     65    0   Placebo    0.000   0.000   0   0        0
#> 15   15   Male  27     61  200      Drug 1362.365 113.832   1   0        1
#> 16   16   Male  24     73  100      Drug  458.437  88.072   1   0        0
#> 17   17   Male  24     71  200      Drug  582.983  43.995   0   0        0
#> 18   18   Male  29     67    0   Placebo    0.000   0.000   0   0        0
#> 19   19   Male  30     68  200      Drug  385.854  37.411   0   0        1
#> 20   20   Male  27     90    0   Placebo    0.000   0.000   0   0        0
#> 21   21 Female  18     78    0   Placebo    0.000   0.000   0   0        0
#> 22   22   Male  35     69  100      Drug 1619.989 178.407   1   1        0
#> 23   23   Male  33     80    0   Placebo    0.000   0.000   0   0        0
#> 24   24 Female  33     62    0   Placebo    0.000   0.000   0   0        1
#> 25   25   Male  25     65  100      Drug 1130.995 116.015   1   0        1
#> 26   26 Female  18     66  100      Drug 1226.539  90.102   1   1        1
#> 27   27   Male  24     77    0   Placebo    0.000   0.000   0   0        0
#> 28   28 Female  25     78    0   Placebo    0.000   0.000   0   0        1
#> 29   29 Female  22     60  200      Drug 1461.410 155.926   1   1        1
#> 30   30 Female  18     55  200      Drug 3255.732 496.641   1   1       14
#> 31   31 Female  30     59  100      Drug  833.211  60.111   1   0        1
#> 32   32   Male  22     72  200      Drug  579.326  57.582   0   0        1
#> 33   33 Female  20     67  200      Drug  428.255  45.684   0   0        2
#> 34   34 Female  18     60  200      Drug 1060.979 114.528   1   1        2
#> 35   35 Female  33     51  100      Drug 1112.251  77.928   1   1        2
#> 36   36   Male  27     72    0   Placebo    0.000   0.000   1   0        1
#> 37   37 Female  25     72  200      Drug  274.341  22.588   1   0        1
#> 38   38 Female  18     65  100      Drug 1364.737  84.569   1   1        1
#> 39   39 Female  20     62  200      Drug  385.888  47.780   0   1        0
#> 40   40   Male  22     64  100      Drug 1455.605 121.905   1   0        3
#> 41   41 Female  28     62    0   Placebo    0.000   0.000   0   0        0
#> 42   42 Female  35     79  200      Drug 2355.535 280.807   1   1        8
#> 43   43   Male  31     81  200      Drug 1122.648 184.408   1   1        2
#> 44   44   Male  27     73  100      Drug  736.412  80.004   1   0        1
#> 45   45 Female  28     63  200      Drug  584.081  40.820   0   0        0
#> 46   46 Female  30     63  100      Drug  869.921  71.521   1   1        2
#> 47   47 Female  32     79  100      Drug 1317.134  60.836   1   1        1
#> 48   48 Female  19     51  100      Drug  383.240  16.196   1   1        0
#> 49   49   Male  24     65    0   Placebo    0.000   0.000   0   0        1
#> 50   50 Female  29     60  200      Drug 2681.917 240.822   1   1        5
#> 51   51 Female  25     64  200      Drug 1887.838 242.097   1   1        3
#> 52   52   Male  27     61    0   Placebo    0.000   0.000   0   0        0
#> 53   53   Male  32     58    0   Placebo    0.000   0.000   0   0        2
#> 54   54 Female  18     74  100      Drug  795.485  66.977   1   1        1
#> 55   55   Male  19     60    0   Placebo    0.000   0.000   0   0        0
#> 56   56   Male  30     77  200      Drug 3722.993 288.252   1   1       12
#> 57   57 Female  31     61  100      Drug  458.388  44.612   1   0        2
#> 58   58   Male  22     75  200      Drug 3052.063 167.613   1   1        9
#> 59   59   Male  28     78  100      Drug  480.918  77.503   1   0        1
#> 60   60 Female  20     57    0   Placebo    0.000   0.000   1   0        2
#> 61   61   Male  27     71  200      Drug 2170.831 262.774   1   1        1
#> 62   62 Female  35     49  100      Drug  789.615  60.099   1   1        1
#> 63   63   Male  18     65  200      Drug 3642.899 373.898   1   1       20
#> 64   64   Male  19     73  100      Drug  354.179  35.209   0   0        1
#> 65   65   Male  27     68    0   Placebo    0.000   0.000   0   0        0
#> 66   66   Male  26     84    0   Placebo    0.000   0.000   0   0        0
#> 67   67 Female  24     64  100      Drug  877.315  96.264   1   0        0
#> 68   68 Female  33     55  100      Drug  525.982  30.491   0   0        1
#> 69   69 Female  28     80    0   Placebo    0.000   0.000   0   0        1
#> 70   70 Female  22     62  200      Drug 1044.014 150.891   1   0        0
#> 71   71   Male  27     72    0   Placebo    0.000   0.000   0   0        1
#> 72   72   Male  21     58  200      Drug 1767.373 186.440   1   1        1
#> 73   73 Female  19     57  100      Drug 1044.311 118.105   1   1        1
#> 74   74 Female  20     79  100      Drug 2416.468 185.787   1   1        4
#> 75   75 Female  25     67  100      Drug  226.233  17.545   0   0        1
#> 76   76   Male  35     94  200      Drug 2015.462 115.511   1   1        3
#> 77   77   Male  24     56    0   Placebo    0.000   0.000   0   0        1
#> 78   78   Male  31     56    0   Placebo    0.000   0.000   0   0        0
#> 79   79   Male  25     88  100      Drug  974.216 147.318   1   1        0
#> 80   80   Male  27     76  100      Drug 1808.755 177.276   1   1        3
#> 81   81   Male  21     82  200      Drug  768.955  56.500   1   1        2
#> 82   82 Female  29     64  200      Drug  597.485  86.135   1   1        2
#> 83   83   Male  31     67  200      Drug 1849.664 129.645   1   1        2
#> 84   84   Male  30     82    0   Placebo    0.000   0.000   0   0        0
#> 85   85 Female  31     75    0   Placebo    0.000   0.000   0   0        0
#> 86   86 Female  30     67    0   Placebo    0.000   0.000   0   0        1
#> 87   87 Female  31     55    0   Placebo    0.000   0.000   1   0        1
#> 88   88   Male  29     76  200      Drug 1217.919  93.031   1   1        1
#> 89   89   Male  34     69    0   Placebo    0.000   0.000   0   1        0
#> 90   90   Male  30     68  100      Drug  702.453  30.618   1   0        1
#> 91   91 Female  20     58  100      Drug 1302.042 208.087   1   1        4
#> 92   92 Female  30     53  200      Drug  446.250  51.123   1   0        0
#> 93   93 Female  21     58  100      Drug  257.873  25.630   1   0        1
#> 94   94   Male  27     74  100      Drug 2246.038 201.544   1   1        4
#> 95   95   Male  26     66    0   Placebo    0.000   0.000   0   0        0
#> 96   96   Male  27     57  200      Drug  294.722  22.513   1   0        1
#> 97   97 Female  35     59    0   Placebo    0.000   0.000   0   0        1
#> 98   98   Male  28     63    0   Placebo    0.000   0.000   0   0        1
#> 99   99 Female  27     56    0   Placebo    0.000   0.000   1   0        1
#> 100 100   Male  26     66  200      Drug  441.222  53.433   1   0        0
#> 101 101   Male  29     77  100      Drug  483.016  40.052   1   0        1
#> 102 102   Male  20     91  200      Drug 1807.899 213.184   1   1        3
#> 103 103 Female  35     68  200      Drug 1408.282  87.972   1   0        0
#> 104 104 Female  32     72  100      Drug  793.385 180.396   1   0        1
#> 105 105 Female  18     47  200      Drug  818.907  45.140   1   1        0
#> 106 106 Female  29     51  100      Drug  162.864  13.093   0   0        1
#> 107 107 Female  30     50  200      Drug 3213.400 167.695   1   1       12
#> 108 108 Female  20     61    0   Placebo    0.000   0.000   0   0        1
#> 109 109   Male  19     83  100      Drug  890.558  65.187   1   0        1
#> 110 110   Male  22     73  100      Drug 2924.265 223.050   1   1       10
#> 111 111 Female  33     63    0   Placebo    0.000   0.000   1   0        0
#> 112 112   Male  22     72  200      Drug 2539.476 204.581   1   1        5
#> 113 113   Male  24     68  200      Drug  730.700  97.333   1   0        1
#> 114 114   Male  25     76  200      Drug  263.588  41.749   0   0        0
#> 115 115   Male  26     72  100      Drug 1425.726 177.811   1   1        2
#> 116 116   Male  23     58    0   Placebo    0.000   0.000   0   0        1
#> 117 117 Female  26     56    0   Placebo    0.000   0.000   0   1        1
#> 118 118   Male  28     73  100      Drug  461.749  31.404   1   0        0
#> 119 119 Female  25     63  200      Drug 2291.942  97.114   1   1        5
#> 120 120 Female  31     50  200      Drug  367.062  27.661   1   1        1
#> 121 121 Female  18     59  200      Drug  339.198  52.292   1   1        1
#> 122 122 Female  32     63    0   Placebo    0.000   0.000   0   0        0
#> 123 123 Female  21     64  100      Drug 2043.240 249.904   1   1        1
#> 124 124 Female  24     56  200      Drug  374.902  18.513   1   0        0
#> 125 125 Female  34     71    0   Placebo    0.000   0.000   0   1        0
#> 126 126   Male  29     76    0   Placebo    0.000   0.000   0   0        1
#> 127 127 Female  21     60  100      Drug  471.187  40.504   0   1        0
#> 128 128 Female  27     63  100      Drug 2950.706 274.014   1   1        9
#> 129 129 Female  31     63  200      Drug 1162.900 130.848   1   1        2
#> 130 130 Female  26     50  100      Drug  530.786  40.702   0   0        0
#> 131 131   Male  21     80  200      Drug 1007.338  97.577   1   0        1
#> 132 132   Male  21     68  100      Drug 1580.996 141.095   1   1        2
#> 133 133   Male  34     75  100      Drug 1157.315 203.295   1   0        0
#> 134 134   Male  32     85  100      Drug  224.627  21.835   1   0        1
#> 135 135   Male  31     59    0   Placebo    0.000   0.000   0   0        0
#> 136 136   Male  26     95  200      Drug  246.665  24.680   0   0        0
#> 137 137 Female  19     59  200      Drug  645.604  98.964   1   1        1
#> 138 138 Female  24     65    0   Placebo    0.000   0.000   0   0        0
#> 139 139   Male  19     68  100      Drug 1100.362  62.231   1   1        1
#> 140 140 Female  25     67    0   Placebo    0.000   0.000   1   0        1
#> 141 141   Male  35     69  200      Drug  628.589  97.985   0   0        1
#> 142 142 Female  21     72    0   Placebo    0.000   0.000   0   0        0
#> 143 143 Female  28     66  100      Drug  289.184  47.884   0   0        3
#> 144 144 Female  18     56    0   Placebo    0.000   0.000   0   0        0
#> 145 145   Male  18     64  200      Drug 1208.343 106.931   1   0        2
#> 146 146 Female  21     61  200      Drug 2775.362 498.597   1   1        7
#> 147 147 Female  28     72  100      Drug 2865.488 271.863   1   1        6
#> 148 148 Female  24     63  200      Drug 1677.625  90.128   1   1        3
#> 149 149   Male  25     61    0   Placebo    0.000   0.000   0   1        1
#> 150 150   Male  19     68  100      Drug 2051.814 148.594   1   1        3
#> 151 151   Male  28     73    0   Placebo    0.000   0.000   0   0        1
#> 152 152 Female  34     54  200      Drug  568.096  43.455   1   0        1
#> 153 153 Female  25     66    0   Placebo    0.000   0.000   0   1        0
#> 154 154   Male  32     61  200      Drug 1064.011 143.380   1   0        0
#> 155 155 Female  21     70  100      Drug  935.133 100.014   1   0        1
#> 156 156   Male  25     63  200      Drug  887.804  71.805   1   1        0
#> 157 157   Male  21     78  100      Drug  609.154 131.319   1   1        1
#> 158 158   Male  29     87  200      Drug 2547.683 359.105   1   1        5
#> 159 159   Male  30     65  200      Drug  560.817  89.201   1   0        1
#> 160 160   Male  28     75  100      Drug  341.378  23.414   0   1        0
#> 161 161 Female  31     51    0   Placebo    0.000   0.000   0   0        0
#> 162 162   Male  23     75    0   Placebo    0.000   0.000   0   0        0
#> 163 163   Male  35     79  100      Drug  674.440  73.279   1   0        0
#> 164 164   Male  26     64  100      Drug 1920.372 212.534   1   1        3
#> 165 165   Male  22     67  200      Drug  191.698  12.439   1   0        0
#> 166 166 Female  20     66  100      Drug  382.546  55.297   0   1        0
#> 167 167 Female  29     57    0   Placebo    0.000   0.000   1   0        0
#> 168 168   Male  25     77  100      Drug  339.439  34.730   1   1        0
#> 169 169 Female  28     60    0   Placebo    0.000   0.000   0   0        0
#> 170 170 Female  19     68    0   Placebo    0.000   0.000   0   1        0
#> 171 171 Female  32     60  200      Drug 1520.071 120.531   1   1        1
#> 172 172 Female  24     75  200      Drug  339.041  29.594   0   0        0
#> 173 173 Female  33     62  100      Drug  893.447 123.083   1   1        1
#> 174 174 Female  33     65  100      Drug  212.546  19.156   0   0        0
#> 175 175 Female  25     54    0   Placebo    0.000   0.000   0   0        1
#> 176 176 Female  24     68  200      Drug 1365.614 104.783   1   1        4
#> 177 177   Male  27     77  100      Drug  866.954 121.070   0   1        2
#> 178 178   Male  24     73  100      Drug 2454.937 197.494   1   1        6
#> 179 179 Female  28     62  200      Drug  497.479  48.541   1   0        0
#> 180 180   Male  24     68  200      Drug 1625.451  88.952   1   1        1
#> 181 181   Male  29     82  100      Drug 2504.667 185.931   1   1        7
#> 182 182   Male  33     74  200      Drug  818.650  87.229   1   1        0
#> 183 183   Male  30     62  100      Drug 1468.695 223.967   1   1        3
#> 184 184 Female  33     56    0   Placebo    0.000   0.000   0   0        0
#> 185 185   Male  31     82  200      Drug  503.625  78.531   1   0        1
#> 186 186 Female  20     60    0   Placebo    0.000   0.000   1   0        2
#> 187 187   Male  19     64    0   Placebo    0.000   0.000   1   0        0
#> 188 188   Male  35     69    0   Placebo    0.000   0.000   0   0        0
#> 189 189 Female  23     48  200      Drug 2280.144  92.721   1   1        4
#> 190 190 Female  22     58  100      Drug  867.796  69.574   1   0        0
#> 191 191 Female  26     66  100      Drug 1094.454 187.507   1   1        0
#> 192 192   Male  20     69  100      Drug  810.095  80.556   1   0        0
#> 193 193   Male  22     57  100      Drug  983.889 124.998   1   0        1
#> 194 194 Female  32     72  200      Drug 3255.372 390.221   1   1       17
#> 195 195 Female  20     56  100      Drug 1443.564 114.093   1   1        0
#> 196 196 Female  25     69    0   Placebo    0.000   0.000   0   0        1
#> 197 197 Female  26     69  100      Drug  540.914  78.511   1   0        0
#> 198 198   Male  22     75    0   Placebo    0.000   0.000   0   0        0
#> 199 199   Male  24     56  200      Drug  856.344  61.137   1   0        2
#> 200 200   Male  28     70  200      Drug  799.223  82.090   1   0        0
#> 201 201 Female  21     61  200      Drug 1753.611  94.003   1   1        4
#> 202 202 Female  28     61  100      Drug 2603.925 141.575   1   1        3
#> 203 203   Male  31     57  200      Drug 2577.044 185.829   1   0        4
#> 204 204 Female  20     68  200      Drug  393.769  49.725   1   0        2
#> 205 205   Male  29     83  200      Drug 2065.235 453.338   1   1        0
#> 206 206 Female  34     56    0   Placebo    0.000   0.000   1   0        1
#> 207 207 Female  30     51    0   Placebo    0.000   0.000   0   0        0
#> 208 208   Male  19     58  200      Drug 1724.400 112.840   1   0        1
#> 209 209 Female  27     58    0   Placebo    0.000   0.000   0   1        2
#> 210 210   Male  34     67    0   Placebo    0.000   0.000   0   0        0
#> 211 211 Female  21     51  200      Drug 1542.168 177.613   1   0        1
#> 212 212 Female  27     62  100      Drug  516.935  36.475   1   0        0
#> 213 213   Male  34     70    0   Placebo    0.000   0.000   0   0        0
#> 214 214 Female  32     75  100      Drug  726.037 136.925   1   0        2
#> 215 215 Female  34     50  200      Drug  670.344  77.258   1   0        0
#> 216 216 Female  24     80    0   Placebo    0.000   0.000   0   0        0
#> 217 217   Male  24     70    0   Placebo    0.000   0.000   0   0        0
#> 218 218   Male  30     70  200      Drug  648.465  68.274   1   0        0
#> 219 219 Female  29     61  100      Drug 1081.962 100.874   1   0        0
#> 220 220   Male  34     72  100      Drug  171.230  16.799   0   0        0
#> 221 221   Male  28     90  200      Drug 1901.774 110.022   1   1        3
#> 222 222   Male  24     57    0   Placebo    0.000   0.000   0   0        0
#> 223 223 Female  27     73  200      Drug 1842.741 306.776   1   1        3
#> 224 224 Female  31     73    0   Placebo    0.000   0.000   0   0        0
#> 225 225   Male  19     91  200      Drug  603.230  45.738   1   1        1
#> 226 226 Female  35     62    0   Placebo    0.000   0.000   1   0        0
#> 227 227 Female  33     64  200      Drug 1966.172 303.531   1   1        2
#> 228 228 Female  34     78    0   Placebo    0.000   0.000   0   0        1
#> 229 229 Female  23     50    0   Placebo    0.000   0.000   1   0        0
#> 230 230   Male  33     61  100      Drug 1218.938  73.687   1   1        0
#> 231 231 Female  21     64  100      Drug  234.197  23.744   0   0        1
#> 232 232   Male  27     85  100      Drug  496.346  68.392   1   0        0
#> 233 233   Male  24     79  200      Drug  303.590  25.241   1   0        1
#> 234 234 Female  23     59  100      Drug  675.518  51.047   1   0        1
#> 235 235   Male  18     79    0   Placebo    0.000   0.000   0   0        0
#> 236 236 Female  27     71  200      Drug  331.588  32.546   1   0        1
#> 237 237   Male  18     63  100      Drug 2273.110 135.724   1   1        3
#> 238 238   Male  18     69    0   Placebo    0.000   0.000   1   0        1
#> 239 239   Male  35     91  200      Drug 1789.799 180.536   1   1        1
#> 240 240   Male  26     72  100      Drug 1840.383 170.893   1   1        3
#> 241 241   Male  22     63  100      Drug  594.205  54.350   1   0        0
#> 242 242 Female  35     58  200      Drug 1338.897 147.447   1   1        0
#> 243 243   Male  21     74    0   Placebo    0.000   0.000   1   0        0
#> 244 244 Female  30     67  100      Drug  352.784  36.611   0   0        1
#> 245 245   Male  26     66  100      Drug  228.970  30.119   0   0        0
#> 246 246   Male  31     80    0   Placebo    0.000   0.000   0   1        1
#> 247 247   Male  30     80  200      Drug 3055.035 238.682   1   1        5
#> 248 248   Male  24     74  100      Drug 1012.230  58.333   1   1        2
#> 249 249   Male  27     85    0   Placebo    0.000   0.000   0   1        0
#> 250 250 Female  23     71    0   Placebo    0.000   0.000   0   0        0
#> 251 251 Female  30     49  200      Drug  601.154  60.651   1   1        0
#> 252 252   Male  26     62  100      Drug  372.557  46.866   1   0        1
#> 253 253   Male  21     86  100      Drug  326.805  15.257   1   0        0
#> 254 254   Male  18     77  200      Drug 2484.888 292.077   1   1        8
#> 255 255   Male  20     66  200      Drug  208.314  20.265   0   1        1
#> 256 256   Male  29     91    0   Placebo    0.000   0.000   0   0        0
#> 257 257   Male  18     78    0   Placebo    0.000   0.000   0   0        2
#> 258 258 Female  24     57    0   Placebo    0.000   0.000   0   0        0
#> 259 259   Male  24     81  100      Drug  286.874  39.781   0   0        0
#> 260 260 Female  32     71    0   Placebo    0.000   0.000   0   0        0
#> 261 261 Female  27     67  100      Drug  685.882  85.324   1   0        3
#> 262 262 Female  34     60  100      Drug  435.479  59.803   0   0        0
#> 263 263   Male  25     65  200      Drug  867.893  51.084   1   0        0
#> 264 264   Male  33     80    0   Placebo    0.000   0.000   0   0        0
#> 265 265   Male  20     76    0   Placebo    0.000   0.000   0   0        0
#> 266 266 Female  21     64  100      Drug 1048.372  67.559   1   0        1
#> 267 267 Female  21     56  200      Drug  446.828  41.096   1   0        0
#> 268 268 Female  30     59  200      Drug  833.280 110.670   1   0        3
#> 269 269 Female  26     76  100      Drug  884.206  78.182   1   0        3
#> 270 270   Male  21     83    0   Placebo    0.000   0.000   0   0        0
#> 271 271   Male  25     73    0   Placebo    0.000   0.000   0   0        1
#> 272 272   Male  21     66  200      Drug 2165.424 301.469   1   1        2
#> 273 273 Female  25     65  200      Drug 1121.834 117.531   1   0        1
#> 274 274 Female  31     67    0   Placebo    0.000   0.000   0   0        0
#> 275 275   Male  31     64  100      Drug  827.687  93.514   1   1        1
#> 276 276   Male  20     64  200      Drug 1265.558 154.554   1   1        1
#> 277 277   Male  18     59    0   Placebo    0.000   0.000   0   0        0
#> 278 278 Female  31     57  200      Drug  653.849 120.501   1   0        1
#> 279 279 Female  28     60  100      Drug  991.645  76.434   1   0        2
#> 280 280 Female  23     62  100      Drug  470.267  86.450   0   0        0
#> 281 281 Female  18     68    0   Placebo    0.000   0.000   0   0        0
#> 282 282 Female  33     74    0   Placebo    0.000   0.000   0   0        0
#> 283 283 Female  21     65  100      Drug 2490.561 242.822   1   1        6
#> 284 284 Female  27     58  200      Drug  273.964  16.926   1   0        0
#> 285 285   Male  18     79  100      Drug  376.129  55.439   1   1        0
#> 286 286 Female  28     70    0   Placebo    0.000   0.000   0   1        0
#> 287 287   Male  21     75  100      Drug 3106.729 360.532   1   1        8
#> 288 288   Male  22     74    0   Placebo    0.000   0.000   0   1        1
#> 289 289   Male  32     57    0   Placebo    0.000   0.000   0   0        1
#> 290 290   Male  24     79    0   Placebo    0.000   0.000   0   0        0
#> 291 291   Male  27     60    0   Placebo    0.000   0.000   1   0        0
#> 292 292 Female  29     65    0   Placebo    0.000   0.000   0   0        0
#> 293 293   Male  27     60  100      Drug  809.501  96.285   1   1        1
#> 294 294   Male  21     74  100      Drug 2584.886 502.810   1   1        1
#> 295 295 Female  31     56    0   Placebo    0.000   0.000   0   0        1
#> 296 296 Female  35     72    0   Placebo    0.000   0.000   0   1        0
#> 297 297 Female  25     66    0   Placebo    0.000   0.000   0   0        0
#> 298 298   Male  35     88    0   Placebo    0.000   0.000   0   0        0
#> 299 299   Male  25     70  200      Drug 1000.850 115.328   1   1        2
#> 300 300 Female  28     51  100      Drug  230.663  31.165   0   0        0
#>     biomarker_change ae_duration
#> 1         1.21689461  12.4023385
#> 2         4.86707165  13.6978409
#> 3        -1.83283026   5.2620230
#> 4         1.90016950   6.6994222
#> 5        -1.00917854   6.1528252
#> 6        -4.97275271  12.8297263
#> 7        -2.52085102   2.4668453
#> 8         1.57558989   9.1333783
#> 9        -0.75682794   9.1964781
#> 10       -1.21561961  10.6308157
#> 11        3.57427203  11.0634886
#> 12       -2.56712432   1.9670020
#> 13        1.27913577   4.3028003
#> 14       -1.23608374  12.1104252
#> 15        1.11822628  12.4903220
#> 16        1.38418355  15.5620922
#> 17       -0.71913792   2.5561696
#> 18       -3.66813438  10.0615376
#> 19       -4.45771909  17.8905827
#> 20       -1.02452220   4.3957877
#> 21       -2.14000520   5.2146509
#> 22        3.16842301   1.9134481
#> 23       -4.50396876   2.0389703
#> 24       -3.26981998   4.7136690
#> 25       -0.32164642   7.5497435
#> 26       -2.10561827   0.2920545
#> 27       -1.72295634   5.4315756
#> 28       -2.28667255   5.3509014
#> 29        1.08809263   4.1668039
#> 30        4.04689961   3.3174464
#> 31       -1.64361291   4.6792955
#> 32        0.24899543   8.2899379
#> 33       -1.55566986   8.3434118
#> 34       -0.71864842  22.8656286
#> 35        2.02926745   5.0465865
#> 36       -0.81499572   7.7832669
#> 37        0.13061085   9.3070779
#> 38        0.78992199  12.2776890
#> 39        1.37495249  17.7707121
#> 40        0.97660645   5.4836572
#> 41       -2.28837767   3.9217421
#> 42        2.42035391   7.8067662
#> 43       -0.88553324   7.6996408
#> 44        0.92764434   0.5674367
#> 45        1.32547293   4.1653344
#> 46       -0.25583707   3.8275767
#> 47        2.98472261   6.8976180
#> 48        0.12810597  14.9170763
#> 49       -3.03917422   4.3342904
#> 50        1.34485268   4.2918735
#> 51        0.68069550  10.6544252
#> 52       -1.64544558   0.8458041
#> 53       -1.17821769   1.2087244
#> 54       -2.32785728  12.0272105
#> 55        0.35449731  12.8284294
#> 56        5.04580280  28.1963406
#> 57        0.72124026  14.5485754
#> 58        3.40187383   7.2931356
#> 59       -0.46804132   8.1647782
#> 60       -1.32929069   5.1650589
#> 61        0.81891952  14.7973971
#> 62       -0.24723696   4.2118454
#> 63        4.70961268  24.1159217
#> 64       -2.31000877   7.3382358
#> 65       -4.42837865   7.8871636
#> 66       -3.61836613   5.8482122
#> 67        0.62406680  15.0324361
#> 68        0.55789227   6.1050818
#> 69       -0.42432473   7.9277063
#> 70        1.75635445   3.2647503
#> 71       -3.98436216   0.7060767
#> 72        1.26212024  14.6549675
#> 73        2.66605346   8.0882135
#> 74        2.23853129  17.1259032
#> 75       -1.88058094  29.8256454
#> 76        1.95116147   4.6600693
#> 77       -2.72507616  12.9509667
#> 78       -5.05937586   6.6433803
#> 79       -0.74249331  20.1956141
#> 80        2.76068254   7.5811517
#> 81        0.30863560   8.3324644
#> 82       -0.83605291  12.5603736
#> 83       -0.22328442   6.5505918
#> 84       -3.55297967   1.0712052
#> 85       -3.51942041   8.6298653
#> 86       -1.68193937  10.0021621
#> 87        0.06950358   6.8484105
#> 88        3.37875480  19.1238063
#> 89       -3.28792903   3.3154427
#> 90       -0.15999997   7.8220692
#> 91       -0.49468421  12.1231626
#> 92        0.04917099   9.4914859
#> 93        0.71403936   8.1272546
#> 94        1.37498132   4.0812323
#> 95       -4.45930612   1.5814588
#> 96        0.80122443   8.1843521
#> 97       -0.44500277   3.1820631
#> 98       -2.89513922   0.5840801
#> 99        1.04871516   5.8748827
#> 100       1.20894239  12.6630278
#> 101      -0.26102667   0.5672816
#> 102       0.86454698  15.0618484
#> 103       0.94589128  23.5875300
#> 104       0.55167463   2.6159680
#> 105       2.20668279  11.5488886
#> 106      -0.51127533   6.5726007
#> 107       2.04872721   8.8860617
#> 108      -2.70648916   4.6599942
#> 109       0.16878730   2.1870869
#> 110       3.47671600  19.4375318
#> 111      -3.21548701   2.5782743
#> 112       3.11259886   5.2698982
#> 113      -1.10823684   9.1676330
#> 114      -0.84604263   6.4392208
#> 115       0.19012212  13.4517879
#> 116       0.16086598   8.0337406
#> 117      -0.72487676   5.8531641
#> 118       0.02303031  10.2977946
#> 119       2.89428243  27.7335425
#> 120      -0.07718506   8.9043890
#> 121       2.22020135   2.3710430
#> 122       0.07641323   3.6041786
#> 123       2.29907932   8.8294978
#> 124       0.88635640  13.5085243
#> 125      -1.62331948   3.4806025
#> 126      -2.44167721   5.8746214
#> 127       0.13185178   4.3649435
#> 128       7.01580871   7.9848829
#> 129      -0.29303233  15.2301126
#> 130      -0.85034937  18.5908812
#> 131       0.98115923   6.5277101
#> 132       3.02248761  20.7351612
#> 133      -0.10222225   5.3984136
#> 134      -1.25664071   6.9658896
#> 135      -3.51164530   7.7420600
#> 136      -2.22622282   8.3692315
#> 137       0.08381277   4.6983414
#> 138      -0.69945716   8.4329764
#> 139       1.54593637   2.8279543
#> 140      -2.54803535   9.4879645
#> 141       2.19673335  15.3569904
#> 142      -0.44345652   1.9759314
#> 143       2.61900069  11.7800952
#> 144      -0.70961242   0.5571403
#> 145      -0.96311254   3.9164444
#> 146       4.10336606  10.2455753
#> 147       3.25270329   3.0181002
#> 148       1.01462027   7.1029847
#> 149      -3.18581267   9.2750965
#> 150       1.41894713   6.2680416
#> 151       0.69931331   0.8964128
#> 152       1.32959945   7.8288020
#> 153      -1.70751950   6.4255340
#> 154       1.99457781  25.8997383
#> 155       1.74924806   3.8293527
#> 156      -1.40155209   8.6630560
#> 157      -0.71748667   3.2437917
#> 158       6.22339571  10.1899600
#> 159       0.38559205  15.1246136
#> 160      -0.22404186   3.5677195
#> 161       0.09884469   3.9378639
#> 162      -0.91481807   2.5622551
#> 163       2.88244386   2.0393986
#> 164       1.71793139   2.1229870
#> 165      -0.32941303  13.4455356
#> 166       0.21905160   6.1791956
#> 167      -0.70123075   6.1290166
#> 168      -0.91415250   2.8938697
#> 169      -0.74040330   2.3392964
#> 170      -2.32276260   2.0888275
#> 171       1.34734543   7.2580183
#> 172      -1.10902365   3.9842946
#> 173       0.04218704  11.0482530
#> 174      -3.83687699  13.5673715
#> 175      -2.80847388   4.9578695
#> 176       0.19536822   9.1565528
#> 177       1.36079568  14.0041916
#> 178       3.50538382  14.7788296
#> 179      -2.42962603   6.5622627
#> 180       1.86555267  29.4220319
#> 181       0.04580699  18.4822193
#> 182      -0.18987328   6.9697865
#> 183      -1.51375403  14.3623507
#> 184       0.25867301   6.8641461
#> 185      -0.58912179   4.5625637
#> 186      -1.19471648   2.3755352
#> 187       0.43548839   3.7421468
#> 188      -2.01472438   6.3518476
#> 189       1.71519107  22.8222778
#> 190       0.86128551   1.3109807
#> 191       3.35901223  11.4208081
#> 192       0.36523138   9.5577864
#> 193       0.60240105   5.7318496
#> 194       3.19048563  25.3474553
#> 195       1.11269082  23.3473527
#> 196      -4.34989961  22.3243441
#> 197       0.10373240   2.4786837
#> 198      -0.85183003   8.1391839
#> 199      -0.21256662   4.2316224
#> 200       1.13530930  14.2056126
#> 201       3.98094297  15.5107844
#> 202       3.62231164   7.3471554
#> 203       4.54384835  16.5507347
#> 204      -0.12008136   5.2956758
#> 205       4.11062841  21.8441725
#> 206      -1.66767305  11.2656664
#> 207      -2.93879916   0.6290950
#> 208       0.32008656  24.2130542
#> 209       1.28604036   2.1066428
#> 210      -3.10133943   4.9605744
#> 211       0.90859042   8.2630988
#> 212      -2.43395176  12.2442650
#> 213      -4.77946861   2.2082908
#> 214       2.18812784   9.3870753
#> 215       1.02002553  25.5800090
#> 216      -2.88518963   6.7570156
#> 217      -3.43323569   2.0388772
#> 218      -3.35209629   1.9681410
#> 219       2.82641088  21.4114995
#> 220      -1.87449081   8.7774109
#> 221       1.10308744   6.8868818
#> 222      -3.20825352   2.8121970
#> 223       4.22401194   8.1476937
#> 224      -2.17028872   2.7697333
#> 225      -0.57323978   5.9615701
#> 226      -0.72178140   3.9845896
#> 227       2.51169211   3.1609063
#> 228      -3.30025414   4.3607234
#> 229      -3.12103828   3.5689390
#> 230       0.41975372  16.3844700
#> 231      -0.71538112  21.3454401
#> 232       0.26098610   7.1373689
#> 233      -1.78702953   5.9261196
#> 234      -0.74311533  30.8523868
#> 235      -1.22635178  17.7333153
#> 236       0.01587594   0.8405244
#> 237       1.13662421   4.4848995
#> 238      -2.24480027   2.4398491
#> 239      -1.49976329  12.0634661
#> 240       1.02464358  14.6713668
#> 241      -0.28612306   3.5806840
#> 242      -0.55613277   7.9954005
#> 243      -3.01876986   3.6217984
#> 244      -1.66120465   8.5610505
#> 245      -0.93444905  18.8411378
#> 246      -1.47604141  11.4968616
#> 247       2.39777540  28.5096658
#> 248      -0.44738016   3.9491263
#> 249      -1.59397744   2.3287287
#> 250      -0.96950648   7.8881917
#> 251      -0.51959465   5.2348218
#> 252      -2.49390578   0.5981972
#> 253      -1.00342300   5.5630495
#> 254       4.06040225  10.2820965
#> 255      -3.90556711  16.0049807
#> 256      -0.43819676   1.5755926
#> 257      -4.49746604   7.4438288
#> 258      -0.82398046   8.0790003
#> 259      -3.30841697   6.2661056
#> 260      -0.11803860   5.3280103
#> 261       0.54389259   6.6116125
#> 262       0.29345923   3.0804305
#> 263      -2.43586295  26.6612423
#> 264      -1.84962144   2.7414151
#> 265      -0.66777861   3.7712780
#> 266       0.18596246  14.6393679
#> 267      -0.89846242   2.2010973
#> 268      -0.31650799   2.3480208
#> 269       0.97401363   5.8607358
#> 270      -1.29085550   2.4623150
#> 271      -1.52548772   3.5544724
#> 272       2.45706590  16.1028356
#> 273       0.30104449   6.5588134
#> 274      -4.86657170   3.4269166
#> 275      -0.80052328  14.1938214
#> 276       1.58466149   9.2418689
#> 277      -2.71544212   1.9708437
#> 278      -2.96415361   7.7012488
#> 279      -1.49068324  11.6991615
#> 280      -4.61016107   2.1487510
#> 281      -2.87202388   8.7470260
#> 282      -0.71731777  13.8842448
#> 283       3.14857930  25.1623546
#> 284       1.72412108   7.3492238
#> 285      -1.53728890   7.2980775
#> 286      -4.07099689   5.7767965
#> 287       2.44007328   6.0093271
#> 288      -3.08328079   4.7560591
#> 289      -1.58995085   3.4651822
#> 290      -0.71932431   3.5004610
#> 291       0.80886967   3.2268671
#> 292      -1.22600448   8.6402485
#> 293      -0.84221263   2.0273475
#> 294       4.79461969   7.9842162
#> 295       1.09619679   4.5882999
#> 296      -2.53869560   3.2174202
#> 297       2.29481706   3.3965829
#> 298      -2.51841551  14.3529439
#> 299      -0.48932292  10.2645511
#> 300      -2.21636531   4.7939398